Worked on basic combat systems
Each agent can be assigned a set of combat abilities, which can be used to deal damage to other agents.
CombatAbility is a scriptable object that can be created in the project and assigned to combatAbilitySets
AI will attack back which ever agent is dealing damage to it.
There is a bug where agents won't attack if they broke an initial attack but returned to attack again.